ZOOKEEPER-4289: Reduce the performance impact of Prometheus metrics
Enabling Prometheus provider has significant impact on the performance of both read and write operations. This is to reduce the impact by making the Prometheus summary reporting as an async operation.

Load test results showed that the avg read latency and throughput after the fix is on par with the performance when Prometheus is disabled. For writes, the avg latency was reduced 25% and the avg throughput was increased 20% after the fix.

Author: liwang <liwang@apple.com>

Reviewers: Enrico Olivelli <eolivelli@apache.org>, Andor Molnar <anmolnar@apache.org>

Closes #1698 from li4wang/ZOOKEEPER-4289

(cherry picked from commit a692cbab92e9bae53c9313241d9668d9de97ed8b)